+++ src/main/java/com/sampleProject/Flynaut/repository/DepartmentRepository.java
@@ -3,5 +3,5 @@ package com.sampleProject.Flynaut.repository;
 import com.sampleProject.Flynaut.entity.department;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DepartmentRepository extends JpaRepository<department, Integer> {
+public interface DepartmentRepository extends JpaRepository<department, Long> {
 }
