+++ src/main/java/com/sampleProject/Flynaut/repository/DesignationRepository.java
@@ -3,5 +3,5 @@ package com.sampleProject.Flynaut.repository;
 import com.sampleProject.Flynaut.entity.designation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DesignationRepository extends JpaRepository<designation, Integer> {
+public interface DesignationRepository extends JpaRepository<designation, Long> {
 }
